+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Music/Play.cs
@@ -7,7 +7,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
 {
     public class Play : KaguyaBase
     {
-        [DisabledCommand]
         [MusicCommand]
         [Command("Play")]
         [Summary("Searches YouTube for the provided song and plays/queues the first " +