+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Music/Promote.cs
@@ -14,7 +14,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
 {
     public class Promote : KaguyaBase
     {
-        [DisabledCommand]
         [MusicCommand]
         [Command("Promote")]
         [Summary("Promotes a track from the queue to the top of the queue. This would " +