+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Music/Queue.cs
@@ -14,7 +14,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
 {
     public class Queue : KaguyaBase
     {
-        [DisabledCommand]
         [MusicCommand]
         [Command("Queue")]
         [Alias("q")]